fix (WMS): do not import DB in utilities, move things in DB.StatusUtils
